delete data file
